Readme
In this project, I learned how to preprocess and merge datasets to calculate needed measures and prepare them for an Analysis. In this project, I worked with the COVID19 datasets, published by John Hopkins University, which consists of the data related to the cumulative number of confirmed cases, per day, in each country. Also, there is another dataset consist of various life factors, scored by the people living in each country around the globe. I merged these two datasets to see if there is any relationship between the spread of the virus in a country and how happy people are, living in that country.